+++ R/is_false.R
@@ -1,0 +1,1 @@
+is_false <- function(x) is.logical(x) && length(x) == 1L && !is.na(x) && !x